requirements.txt
aws_lambda_powertools==2.32.0
boto3==1.34.27
botocore==1.34.27
requests==2.31.0
pydantic==2.5.3
python-jose==3.3.0